chore: Replace type-only export with full export
Type-only exports are only supported in Typescript 3.8 or higher. The type-only export prevents consumers of the package to use it in an environment that is using an older Typescript version.

This replaces the type-only export with a full export. This should not make any difference in the generated JS output of the internal folder's index.ts file.